--- com.unity.cinemachine/Runtime/Helpers/UpgradeFromCM2toCM3.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b5534da2d85f74466947156519ca18d9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
