+++ com.unity.cinemachine/Runtime/Upgrader.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 67c7396a4e6a403e8628a7c83ffd74dd
+timeCreated: 1652452751
