+++ com.unity.cinemachine/Runtime/Upgrader/CinemachineUpgraderDriver.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: e6600a8ee33540888904c7caa4db93d7
+timeCreated: 1652454614
